Start implementation of product CRUD actions

+++ AugustaGourmet.Api.Application/Features/Products/DeleteProduct/DeleteProductCommandHandler.cs
@@ -11,12 +11,14 @@ namespace AugustaGourmet.Api.Application.Features.Products.DeleteProduct;
 public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand, ErrorOr<Unit>>
 {
     private readonly IProductRepository _productRepository;
+    private readonly IInventoryParameterRepository _inventoryParameterRepository;
     private readonly IUnitOfWork _unitOfWork;
 
-    public DeleteProductCommandHandler(IProductRepository productRepository, IUnitOfWork unitOfWork)
+    public DeleteProductCommandHandler(IProductRepository productRepository, IUnitOfWork unitOfWork, IInventoryParameterRepository inventoryParameterRepository)
     {
         _productRepository = productRepository;
         _unitOfWork = unitOfWork;
+        _inventoryParameterRepository = inventoryParameterRepository;
     }
 
     public async Task<ErrorOr<Unit>> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
@@ -27,6 +29,8 @@ public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand,
             return Errors.CouldNotFind(nameof(Product), request.Id);
 
         // validate relationships
+        if (await _inventoryParameterRepository.ProductIsMapped(request.Id))
+            return Errors.Products.Conflicts.WithInventoryParameters;
 
         // Delete from database
         _productRepository.Delete(productToDelete);
